Refactoring Logo and Main
