+++ project/src/components/logo/logo.tsx
@@ -1,6 +1,8 @@
 import { AppRoute } from 'const';
 import { Link, useLocation } from 'react-router-dom';
 
+const LogoImage = () => (<img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />);
+
 function Logo(): JSX.Element {
   const { pathname } = useLocation();
 
@@ -8,13 +10,13 @@ function Logo(): JSX.Element {
     case (AppRoute.Main):
       return (
         <div className="header__logo-link">
-          <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
+          <LogoImage />
         </div>
       );
     default:
       return (
         <Link className="header__logo-link" to="/">
-          <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
+          <LogoImage />
         </Link>
       );
   }
